+++ app/src/main/java/felixsam/github/com/foodordering/activities/List_Orders_Activity.java
@@ -63,7 +63,7 @@ public class List_Orders_Activity extends AppCompatActivity {
                 Order order = new Order(data.getString(data.getColumnIndex(DatabaseHelper.ITEMS_COL3_FIRST_NAME))
                         , data.getString(data.getColumnIndex(DatabaseHelper.ITEMS_COL5_ITEM_NAME))
                         , data.getInt(data.getColumnIndex(DatabaseHelper.ITEMS_COL6_PRICE))
-                        , data.getInt(data.getColumnIndex("TOTAL_ITEMS_PRICE"))
+                        , data.getDouble(data.getColumnIndex("TOTAL_ITEMS_PRICE"))
                         , date
                         , OrderID
                         , data.getInt(data.getColumnIndex(DatabaseHelper.ITEMS_COL7_QUANTITY))
